+++ client/store/auth/actions.js
@@ -1,4 +1,5 @@
 import axios from 'axios'
+import { clearAllSetTimouts } from '../setTImeouts/action'
 
 // --------------------------------------------------
 // Action types
@@ -44,6 +45,7 @@ export const logoutUserThunk = history => {
     return axios
       .delete('/api/auth')
       .then(() => dispatch(logutUser()))
+      .then(() => dispatch(clearAllSetTimouts()))
       .then(() => history.push('/'))
   }
 }
