+++ client/store/socket/Socket.js
@@ -34,10 +34,6 @@ class Socket {
         )
       )
     })
-
-    this.socket.on('stretch-closed', cohortStretch => {
-      storeAPI.dispatch(updateCohortStretch(cohortStretch.id, cohortStretch))
-    })
   }
 
   generateFlashMessageObjectForOpenStretch = cohortStretch => {
@@ -94,10 +90,6 @@ class Socket {
 
   startStretchTimer(cohortStretch) {
     this.socket.emit('startStretchTimer', cohortStretch)
-  }
-
-  sendClosedStretch(cohortStretch) {
-    this.socket.emit('sendClosedStretch', cohortStretch)
   }
 }
